Attempt to address review comments
