+++ tst/scripts/ssheet/ssheet.py
@@ -26,6 +26,7 @@ from scipy.interpolate import interp1d
 logger = logging.getLogger("artemis" + __name__[7:])  # set logger name
 logging.getLogger("h5py").setLevel(logging.WARNING)
 logging.getLogger("matplotlib").setLevel(logging.WARNING)
+import scripts.utils.analysis as analysis
 import matplotlib.colors as colors
 import matplotlib.pyplot as plt
 
@@ -49,7 +50,7 @@ def analyze():
     os.makedirs(artemis.get_fig_dir(), exist_ok=True)
     analyze_status = True
 
-    time, x, y, z, [d, u, v, w, T] = artemis.load_level(
+    time, x, y, z, [d, u, v, w, T] = analysis.load_level(
         "final", dir=artemis.get_data_dir(), base="{}.out1".format(_file_id)
     )
     xc = 0.5 * (x[1:] + x[:-1])
@@ -102,7 +103,7 @@ def analyze():
     axes[1].legend(loc="best", fontsize=12)
     axes[1].set_ylabel("$\\Sigma - \\langle \\Sigma \\rangle$", fontsize=20)
     axes[0].set_ylabel("$y$", fontsize=20)
-    artemis.create_colorbar(axes[0], norm=norm)
+    analysis.create_colorbar(axes[0], norm=norm)
     fig.tight_layout()
     fig.savefig(
         os.path.join(artemis.get_fig_dir(), _file_id + "_spiral.png"),
